requirements.txt
# J.A.R.V.I.S — Complete Dependencies

# ─── Core ML & RL ─────────────────────────────────────────────────────────────
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# ─── Memory & Storage ─────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ─── LLM Providers (install the ones you use) ─────────────────────────────────
google-generativeai>=0.8.0
openai>=1.59.0
anthropic>=0.42.0

# ─── Voice / Speech ───────────────────────────────────────────────────────────
edge-tts>=6.1.9
SpeechRecognition>=3.10.0
pygame>=2.6.0

# ─── Web Dashboard ────────────────────────────────────────────────────────────
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# ─── Async & Scheduling ───────────────────────────────────────────────────────
apscheduler>=3.10.0

# ─── System & Monitoring ──────────────────────────────────────────────────────
psutil>=5.9.0
requests>=2.31.0

# ─── Config & Environment ─────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.0
rich>=13.7.0
colorama>=0.4.6

# ─── Logging ──────────────────────────────────────────────────────────────────
loguru>=0.7.2

# ─── Optional: Semantic Memory ────────────────────────────────────────────────
sentence-transformers>=2.2.2

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
